rename ping service file
