+++ src/lib/server/services.ts
@@ -1,5 +1,5 @@
 import type {Service} from '$lib/server/service';
-import {pingService} from '$lib/server/serverServices';
+import {pingService} from '$lib/server/pingService';
 import {createPersonaService} from '$lib/vocab/persona/personaServices';
 import {
 	readCommunityService,
